passList and Error Handling
Added passList as a function to counter only being able to call the list once. Added Error Handling. I'm sure there's a better way to do it, but I couldn't figure it out. Therefore, I added the primary while loop to a try/except. I then had to add that try/except to another while loop so that the program would not close after 'except' triggers.